+++ cpp/image.cpp
@@ -13,10 +13,10 @@ void implement_image(jlcxx::Module& module)
     module.add_type(Image)
     .add_constructor()
     .add_constructor(const std::string&)
-    .constructor([](size_t width, size_t height, jl_value_t* rgba){
+    .constructor([](uint64_t width, uint64_t height, jl_value_t* rgba){
         return new Image(width, height, unbox_rgba(rgba));
     }, USE_FINALIZERS)
-    .method("create!", [](Image& image, size_t width, size_t height, jl_value_t* rgba){
+    .method("create!", [](Image& image, uint64_t width, uint64_t height, jl_value_t* rgba){
         image.create(width, height, unbox_rgba(rgba));
     })
     .add_type_method(Image, create_from_file, !)
@@ -29,13 +29,13 @@ void implement_image(jlcxx::Module& module)
     .add_type_method(Image, as_scaled)
     .add_type_method(Image, as_cropped)
     .add_type_method(Image, as_flipped)
-    .method("set_pixel_rgba!", [](Image& image, size_t x, size_t y, jl_value_t* rgba){
+    .method("set_pixel_rgba!", [](Image& image, uint64_t x, uint64_t y, jl_value_t* rgba){
         image.set_pixel(x, y, unbox_rgba(rgba));
     })
-    .method("set_pixel_hsva!", [](Image& image, size_t x, size_t y, jl_value_t* hsva){
+    .method("set_pixel_hsva!", [](Image& image, uint64_t x, uint64_t y, jl_value_t* hsva){
         image.set_pixel(x, y, unbox_hsva(hsva));
     })
-    .method("get_pixel", [](const Image& image, size_t x, size_t y) -> jl_value_t* {
+    .method("get_pixel", [](const Image& image, uint64_t x, uint64_t y) -> jl_value_t* {
         return box_rgba(image.get_pixel(x, y));
     })
     ;
